friendship ended with sqlite3. tinydb is my new best friend
