+++ pyproject.toml
@@ -13,6 +13,7 @@ es8 = "bulk_editor.main:app"
 python = "^3.10"
 typer = {extras = ["all"], version = "^0.7.0"}
 asciimatics = "^1.14.0"
+tinydb = "^4.7.1"
 
 
 [tool.poetry.group.dev.dependencies]
